ubuntu_requirements.txt
absl-py==0.14.0
astor==0.8.1
bert-serving-server==1.10.0
cached-property==1.5.2
dataclasses==0.8
gast==0.5.2
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.40.0
h5py==3.1.0
importlib-metadata==4.8.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
pkg-resources==0.0.0
protobuf==3.18.0
pyzmq==22.3.0
six==1.16.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.9.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
